requirements.txt
pycryptodome==3.19.0
PyQt5==5.15.9 